feat(electron): use extra module path provided by electron
When the native sqlite dep is built by electron, it gets placed in

opensphere-electron/app/node_modules, and opensphere-electron provides an environment variable with

the location. Use that variable to prepend that path onto module.paths so that native deps are

loaded from that location in dev, rather than relying on whatever was downloaded by yarn/npm to the

project/hoisted node_modules.